requirements.txt
aiohttp==3.12.*
asgiref==3.9.*
celery[redis]
certifi==2025.*
crispy-bootstrap5==2025.*
Django==5.2.*
#django-allauth==65.7.*
django-bootstrap5==25.*
django-classy-tags==4.1.*
django-cors-headers==4.7.*
django-debug-toolbar==5.2.*
django-environ==0.12.*
django-ratelimit==4.1.*
django-resized==1.0.*
django-sekizai==4.1.*
django-ninja==1.4.*
djangorestframework-api-key==3.1.*
hypercorn==0.17.*
langcodes==3.5.*
language_data==1.3.*
markdown==3.*
pillow==11.3.*
psycopg==3.2.*
python-dateutil==2.9.*
python-dotenv==1.1.*
pytz==2025.*
redis==6.4.*
requests==2.*
sentry-sdk==2.*
rsa==4.9.*
tzdata==2025.*
uritemplate==4.1.*
urllib3>=2.5.0
wheel==0.45.*
whitenoise==6.9.*
yarl==1.20.*